Refactored intro and fixed stacks for building comparison
