+++ energy-app-react/App.js
@@ -15,11 +15,8 @@ import CurrTheme from './src/styling/CurrentTheme'
 import { handler, dataReducer, layoutReducer } from './src/helpers/ReduxHandler'
 import { getCurrentGenerationGraphFormat, getCurrentConsumptionGraphFormat } from './src/helpers/ApiWrappers';
 import SustainStack from './src/SustainView';
-import IntroSlider from './src/IntroSlider';
-import { checkIfFirstLaunch } from './src/checkIfFirstLaunch';
-
-// const defaultFont = CurrFont+'-regular';
-// const defaultFontBold = CurrFont+'-bold';
+import IntroSlider from './src/intro/IntroSlider';
+import { checkIfFirstLaunch } from './src/intro/checkIfFirstLaunch';
 
 const apiGoogleKey = 'AIzaSyA2Q45_33Ot6Jr4EExQhVByJGkucecadyI';
 const themeStyles = GetStyle();
