+++ energy-app-react/src/BuildingListView.js
@@ -6,102 +6,123 @@ import { List, Card, ListItem, Button, Avatar, Header } from 'react-native-eleme
 import { Table, TableWrapper, Row, Rows, Col, Cols, Cell } from 'react-native-table-component';
 import buildings from './Buildings';
 import IndividualBuilding from './IndividualBuilding';
+import ComparisonPage from './ComparisonPage';
 import BuildingComparison from './BuildingComparison';
-import OverviewCards from './overview/OverviewCards';
+// import SelectStack from './IndividualBuilding';
 import { getCurrentBuildingUtilityConsumption } from './helpers/ApiWrappers.js';
-import  ComparisonPage from './ComparisonPage';
 
 class BuildingListView extends Component {
+  renderHeader = (headerItem) => {
+    return <Text style={styles.header}>{headerItem.section.name}</Text>
+  }
 
-    renderHeader = (headerItem) => {
-        return <Text style={styles.header}>{headerItem.section.name}</Text>
-    }
+  renderItem = (item) => {
+      return <View style={{flex: 1, flexDirection: 'row', justifyContent: 'space-between', borderBottomColor: 'black',
+      borderBottomWidth: 1}}>
+          <Image
+          style={{alignItems:'center', width:60, paddingBottom:"10%"}} source={{uri: item.item.avatar}}/>
+          <View style={{flex: 1, flexDirection: 'column', paddingTop:'2%'}}>
+              <Text style={styles.text}>Electricity: {item.item.electricity}</Text>
+              <Text style={styles.text}>Water: {item.item.water}</Text>
+              <Text style={styles.text}>Heat: {item.item.heat}</Text>
+          </View>
+          <Button
+              rightIcon={{name: "angle-right", type: 'font-awesome', size: 20}}
+              fontSize={14}
+              title='More Info'
+              style={{paddingBottom:20}}
+              containerViewStyle={styles.button}
+              backgroundColor='#0B5091'
+              onPress={() => this.props.navigation.navigate('BuildingCardView', {item:item.item})}/>
+      </View>
+  }
 
-    renderItem = (item) => {
-        return <View style={{flex: 1, flexDirection: 'row', justifyContent: 'space-between', borderBottomColor: 'black',
-        borderBottomWidth: 1}}>
-            <Image
-            style={{alignItems:'center', width:60, paddingBottom:"10%"}} source={{uri: item.item.avatar}}/>
-            <View style={{flex: 1, flexDirection: 'column', paddingTop:'2%'}}>
-                <Text style={styles.text}>Electricity: {item.item.electricity}</Text>
-                <Text style={styles.text}>Water: {item.item.water}</Text>
-                <Text style={styles.text}>Heat: {item.item.heat}</Text>
-            </View>
-            <Button
-                rightIcon={{name: "angle-right", type: 'font-awesome', size: 20}}
-                fontSize={14}
-                title='More Info'
-                style={{paddingBottom:20}}
-                containerViewStyle={styles.button}
-                backgroundColor='#0B5091'
-                onPress={() => this.props.navigation.navigate('CardView', {item:item.item})}/>
-        </View>
-    }
+  render() {
+      const {navigate} = this.props.navigation;
 
-    render() {
-        const {navigate} = this.props.navigation;
-
-        return (
-            <List>
-                <SectionList
-                    sections = {buildings}
-                    renderSectionHeader={this.renderHeader}
-                    renderItem={this.renderItem}
-                    keyExtractor = {(item) => item.name}
-                />
-            </List>
-       );
-    }
+      return (
+          <List>
+              <SectionList
+                  sections = {buildings}
+                  renderSectionHeader={this.renderHeader}
+                  renderItem={this.renderItem}
+                  keyExtractor = {(item) => item.name}
+              />
+          </List>
+     );
+  }
 }
 
 
 const BuildingStack = StackNavigator({
-    Buildings: {
-        screen: BuildingListView,
-        navigationOptions: ({ navigation }) => ({
-            title: 'Buildings',
-            ...Platform.select({
-                android: { header: null }
-            }),
-            headerTintColor: 'white',
-            headerStyle: navStyles.header,
-        }),
-    },
-
-    CardView: {
-      screen: IndividualBuilding,
-      path: 'buildings/:name',
+  Buildings: {
+      screen: BuildingListView,
       navigationOptions: ({ navigation }) => ({
-              title: `${navigation.state.params.item.name}`,
-              headerTintColor: 'white',
-              headerStyle: navStyles.header,
-              headerTitleStyle: navStyles.headerTitle,
-              headerBackTitleStyle: navStyles.headerTitle,
-              headerBackTitle: 'Back',
-            }), 
-    },
-    Comparison: {
-        screen: BuildingComparison,
-        navigationOptions: ({ navigation }) => ({
-            title: 'Building Comparison',
-            ...Platform.select({
-                android: { header: null }
-            }),
-            headerTintColor: 'white',
-            headerStyle: navStyles.header,
+          title: 'Buildings',
+          ...Platform.select({
+              android: { header: null }
+          }),
+          headerTintColor: 'white',
+          headerStyle: navStyles.header,
+      }),
+  },
+  // BuildingCardView: {
+  //   screen: IndividualBuilding,
+  //   path: 'buildings/:name',
+  //   navigationOptions: ({ navigation }) => ({
+  //           title: `${navigation.state.params.item.name}`,
+  //           headerTintColor: 'white',
+  //           headerStyle: navStyles.header,
+  //           headerTitleStyle: navStyles.headerTitle,
+  //           headerBackTitleStyle: navStyles.headerTitle,
+  //           headerBackTitle: 'Back',
+  //         }), 
+  // },
+  BuildingCardView: {
+    screen: IndividualBuilding,
+    path: 'buildings/:name',
+    navigationOptions: ({ navigation }) => ({
+      title: `${navigation.state.params.item.name}`,
+      headerTintColor: 'white',
+      headerStyle: navStyles.header,
+      headerTitleStyle: navStyles.headerTitle,
+      headerBackTitleStyle: navStyles.headerTitle,
+      headerBackTitle: 'Back',
+    }), 
+  },
+  // Comparison: {
+  //   screen: BuildingComparison,
+  //   navigationOptions: ({ navigation }) => ({
+  //     title: 'Building Comparison',
+  //     ...Platform.select({
+  //         android: { header: null }
+  //         ios
+  //     }),
+  //     headerTintColor: 'white',
+  //     headerStyle: navStyles.header,
+  //   }),
+  // },
+  Comparison: {
+    screen: BuildingComparison,
+    navigationOptions: ({ navigation }) => ({
+      ...Platform.select({
+          android: { header: null }
+      }),
+      headerTintColor: 'white',
+      headerStyle: navStyles.header,
+    }),
+  },
+  ComparisonPage: {
+    screen: ComparisonPage,
+    navigationOptions: ({ navigation }) => ({
+        title: 'Comparison',
+        ...Platform.select({
+            android: { header: null }
         }),
-    },
-    ComparisonPage: {
-        screen: ComparisonPage,
-        navigationOptions: ({ navigation }) => ({
-            title: 'Compare',
-            ...Platform.select({
-                android: { header: null }
-            }),
-            headerTintColor: 'white',
-            headerStyle: navStyles.header,
-        }),
-    },
+        headerTintColor: 'white',
+        headerStyle: navStyles.header,
+    }),
+  },
 });
 
 const navStyles = StyleSheet.create({
@@ -110,15 +131,6 @@ const navStyles = StyleSheet.create({
     },
 })
 
-
-// styles.listItem
-// styles.text
-// styles.subtitleText
-// styles.listImg
-// navStyles.header
-// navStyles.header
-// navStyles.headerTitle
-// navStyles.headerTitle
 
 const styles = StyleSheet.create({
   card: {
