+++ energy-app-react/src/intro/IntroSlider.js
