+++ energy-app-react/src/intro/checkIfFirstLaunch.js
